# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Stargate API
  description: |
    Stargate is a high-performance API gateway that provides comprehensive route management, 
    load balancing, plugin system, and developer portal functionality.
    
    ## Authentication
    
    Stargate supports multiple authentication methods:
    - **API Key**: Use `X-Admin-Key` header for admin operations
    - **JWT Bearer**: Use `Authorization: Bearer <token>` for user operations
    
    ## Rate Limiting
    
    API requests are subject to rate limiting:
    - Admin API: 1000 requests/hour per API key
    - Portal API: 100 requests/minute per user
    
    ## Error Handling
    
    All endpoints return consistent error responses with HTTP status codes and JSON error objects.
  version: 1.0.0
  contact:
    name: Stargate Team
    url: https://github.com/stargate/stargate
    email: support@stargate.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090
    description: Development server (Admin API)
  - url: http://localhost:8080
    description: Development server (Portal API)
  - url: https://api.stargate.io
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/api-keys:
    post:
      summary: Generate API key
      description: Generate a new API key for authentication
      operationId: generateApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyRequest'
      responses:
        '200':
          description: API key generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    ApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the API key
          example: my-api-key
    ApiKeyResponse:
      type: object
      properties:
        api_key:
          type: string
          description: Generated API key
          example: sk-1234567890abcdef
        name:
          type: string
          description: Name of the API key
          example: my-api-key
        created:
          type: integer
          format: int64
          description: Unix timestamp of creation
          example: 1640995200
...
    Error:
      type: object
      required:
        - error
        - status
      properties:
        error:
          type: string
          description: Error message
          example: Validation failed
        status:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: string
          description: Additional error details
          example: Field 'name' is required
        code:
          type: string
          description: Error code for programmatic handling
          example: VALIDATION_ERROR
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the error
          example: 1640995200
...